Interactive Debug added
GLDL is now compiled as a static library. I had to split gldl in
two distinct files.
gldl_funcarray define an array of all GL functions in string form
and in sorted alphabetical order. This array is used during debug
to set and delete breakpoints.

Breakpoints are functionnal.
However, I had to add a gldlTerminate function to close the lib and
free allocated space, and flush the log.