Added Texture visualisation
With 'pt' or 'printtexture', the texture id given in param is displayed
in a new window.
I need to make texture array to keep track of created textures and their
dimensions, like I did with buffers and shaders.

Pointers given as glFuncs parameters are now displayed as addresses in
the logs.